% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{barplot_density_allen_coral_image}
\alias{barplot_density_allen_coral_image}
\title{Barplot of mean density per coral habitat with data Allen with image}
\usage{
barplot_density_allen_coral_image(df, species, img)
}
\arguments{
\item{img}{}
}
\value{

}
\description{
Barplot of mean density per coral habitat with data Allen with image
}
