% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{barplot_density_allen_coral_mpa}
\alias{barplot_density_allen_coral_mpa}
\title{Barplot of mean density per coral habitat and mpa status}
\usage{
barplot_density_allen_coral_mpa(df, species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Barplot of mean density per coral habitat and mpa status
}
