% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{barplot_density_allen_coral_per_date}
\alias{barplot_density_allen_coral_per_date}
\title{Barplot of mean density in coral habitat per date with data Allen}
\usage{
barplot_density_allen_coral_per_date(df, species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Barplot of mean density in coral habitat per date with data Allen
}
