% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{barplot_density_coral}
\alias{barplot_density_coral}
\title{Barplot of mean density per coral habitat}
\usage{
barplot_density_coral(df, coral_attrib, species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Barplot of mean density per coral habitat
}
