% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{barplot_density_mpa}
\alias{barplot_density_mpa}
\title{Barplot of mean density per mpa status (in/out mpa)}
\usage{
barplot_density_mpa(df, species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Barplot of mean density per mpa status (in/out mpa)
}
