% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{barplot_density_se_mpa_image}
\alias{barplot_density_se_mpa_image}
\title{Barplot of mean density with SE per mpa status (in/out mpa) with image}
\usage{
barplot_density_se_mpa_image(df, species, img)
}
\arguments{
\item{img}{}
}
\value{

}
\description{
Barplot of mean density with SE per mpa status (in/out mpa) with image
}
