% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{barplot_habitat_proportion}
\alias{barplot_habitat_proportion}
\title{barplot habitat proportion in whole sampled area}
\usage{
barplot_habitat_proportion(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
barplot habitat proportion in whole sampled area
}
