% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{barplot_habitat_proportion_in_out_mpa}
\alias{barplot_habitat_proportion_in_out_mpa}
\title{barplot habitat proportion in and out of mpa}
\usage{
barplot_habitat_proportion_in_out_mpa(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
barplot habitat proportion in and out of mpa
}
