% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{boxplot_density_allen_coral}
\alias{boxplot_density_allen_coral}
\title{Boxplot of mean density per coral habitat (0s removed) with data Allen}
\usage{
boxplot_density_allen_coral(df, species)
}
\description{
Boxplot of mean density per coral habitat (0s removed) with data Allen
}
