% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{boxplot_density_allen_coral_mpa_with_megafauna_image_log}
\alias{boxplot_density_allen_coral_mpa_with_megafauna_image_log}
\title{Boxplot of mean density per mpa status and coral habitat (0s removed) with data Allen and megafauna image LOG}
\usage{
boxplot_density_allen_coral_mpa_with_megafauna_image_log(df, species, img)
}
\description{
Boxplot of mean density per mpa status and coral habitat (0s removed) with data Allen and megafauna image LOG
}
