% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{boxplot_density_coral}
\alias{boxplot_density_coral}
\title{Boxplot of mean density per coral habitat (0s removed)}
\usage{
boxplot_density_coral(df, species, coral_attrib)
}
\description{
Boxplot of mean density per coral habitat (0s removed)
}
