% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{boxplot_density_coral_mpa}
\alias{boxplot_density_coral_mpa}
\title{Boxplot of mean density per mpa status and coral habitat (0s removed)}
\usage{
boxplot_density_coral_mpa(df, species, coral_attrib)
}
\description{
Boxplot of mean density per mpa status and coral habitat (0s removed)
}
