% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{boxplot_density_mpa}
\alias{boxplot_density_mpa}
\title{Boxplot of mean density per mpa status (in/out mpa) (0s removed)}
\usage{
boxplot_density_mpa(df, species, img)
}
\description{
Boxplot of mean density per mpa status (in/out mpa) (0s removed)
}
