% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{clean_duplicated_obs_telemetry}
\alias{clean_duplicated_obs_telemetry}
\title{Clean duplicated observations in telemetry}
\usage{
clean_duplicated_obs_telemetry(telem, overl, im_height)
}
\arguments{
\item{telem}{}
}
\value{

}
\description{
Clean duplicated observations in telemetry
}
