% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{clean_off_effort_loop_poe}
\alias{clean_off_effort_loop_poe}
\title{Clean off effort portion related to loops Poe}
\usage{
clean_off_effort_loop_poe(off)
}
\arguments{
\item{off}{}
}
\value{

}
\description{
Clean off effort portion related to loops Poe
}
