% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{compare_densities_between_allen_habitats}
\alias{compare_densities_between_allen_habitats}
\title{Test hyp that median densities between habitat types are different with data Allen}
\usage{
compare_densities_between_allen_habitats(df, species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Test hyp that median densities between habitat types are different with data Allen
}
