% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{compare_densities_between_combined_allen_habitats_mpa_status}
\alias{compare_densities_between_combined_allen_habitats_mpa_status}
\title{Test hyp that median densities between habitat types combined with mpa statusare different with data Allen}
\usage{
compare_densities_between_combined_allen_habitats_mpa_status(df, species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Test hyp that median densities between habitat types combined with mpa statusare different with data Allen
}
