% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{compare_densities_in_out_mpa}
\alias{compare_densities_in_out_mpa}
\title{Test hyp that median densities in mpa and outside mpa are different}
\usage{
compare_densities_in_out_mpa(df, species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Test hyp that median densities in mpa and outside mpa are different
}
