% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{compare_densities_in_out_mpa_greater}
\alias{compare_densities_in_out_mpa_greater}
\title{Test hyp that median densitiy in mpa is greater than outside mpa}
\usage{
compare_densities_in_out_mpa_greater(df, species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Test hyp that median densitiy in mpa is greater than outside mpa
}
