% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_maps.R
\name{convert_telemetry_points_to_lines}
\alias{convert_telemetry_points_to_lines}
\title{convert telemetry points to lines}
\usage{
convert_telemetry_points_to_lines(telem)
}
\arguments{
\item{telem}{}
}
\value{

}
\description{
convert telemetry points to lines
}
