% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{convert_telemetry_spatial}
\alias{convert_telemetry_spatial}
\title{Convert telemetry to spatialpoints dataframe}
\usage{
convert_telemetry_spatial(telem)
}
\arguments{
\item{telem}{}
}
\value{

}
\description{
Convert telemetry to spatialpoints dataframe
}
