% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_maps.R
\name{count_obs_per_grid_per_date}
\alias{count_obs_per_grid_per_date}
\title{Count total number of observations in grid cells per date}
\usage{
count_obs_per_grid_per_date(polygon, telem_obs)
}
\arguments{
\item{telem_obs}{}
}
\value{

}
\description{
Count total number of observations in grid cells per date
}
