% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{create_buffer_transects}
\alias{create_buffer_transects}
\title{create buffer of given width in meters around transect lines}
\usage{
create_buffer_transects(lns, buf_width)
}
\arguments{
\item{buf_width}{}
}
\value{

}
\description{
create buffer of given width in meters around transect lines
}
