% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{intersect_df_coral_poly}
\alias{intersect_df_coral_poly}
\title{Intersect desnity dataframe with coral polygon (ie intersection of points with polygons)}
\usage{
intersect_df_coral_poly(df, coralpoly)
}
\arguments{
\item{coralpoly}{}
}
\value{

}
\description{
Intersect desnity dataframe with coral polygon (ie intersection of points with polygons)
}
