% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{join_video_info_telem}
\alias{join_video_info_telem}
\title{Join video information to telemetry}
\usage{
join_video_info_telem(vid, telem)
}
\value{

}
\description{
Join video information to telemetry
}
