% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{list_poe_videos}
\alias{list_poe_videos}
\title{Get list of on-effort Poe videos}
\usage{
list_poe_videos(vid)
}
\arguments{
\item{vid}{}
}
\value{

}
\description{
Get list of on-effort Poe videos
}
