% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{lmperm_test_habitat_mpa}
\alias{lmperm_test_habitat_mpa}
\title{Barplot with all species : LMperm}
\usage{
lmperm_test_habitat_mpa(df, species1, species2, species3, species4, species5)
}
\arguments{
\item{species5}{}
}
\value{

}
\description{
Barplot with all species : LMperm
}
