% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{make_area_raster}
\alias{make_area_raster}
\title{Make raster for study area with given resolution}
\usage{
make_area_raster(lat1, lon1, lat2, lon2, res)
}
\arguments{
\item{res}{}
}
\value{

}
\description{
Make raster for study area with given resolution
}
