% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{make_coral_cover_raster}
\alias{make_coral_cover_raster}
\title{Make coral cover raster}
\usage{
make_coral_cover_raster(cor, rast)
}
\arguments{
\item{rast}{}
}
\value{

}
\description{
Make coral cover raster
}
