% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{make_df_all_species_poe}
\alias{make_df_all_species_poe}
\title{Make dataframe per grid cell (incl. empty cells) of observations, track length and densities per date and per species for POe}
\usage{
make_df_all_species_poe(polyobs, polytracks, footprintwidth)
}
\arguments{
\item{footprintwidth}{}
}
\value{

}
\description{
Make dataframe per grid cell (incl. empty cells) of observations, track length and densities per date and per species for POe
}
