% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{make_pairwise_comparison_allen_coral_mpa}
\alias{make_pairwise_comparison_allen_coral_mpa}
\title{Make pairwise comparisons for density per coral habitat and mpa}
\usage{
make_pairwise_comparison_allen_coral_mpa(df, species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Make pairwise comparisons for density per coral habitat and mpa
}
