% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{make_permanova_barplot_allen_coral_mpa}
\alias{make_permanova_barplot_allen_coral_mpa}
\title{Make permanova for density per mpa status and coral habitat and barplot of the result with data Allen}
\usage{
make_permanova_barplot_allen_coral_mpa(df, species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Make permanova for density per mpa status and coral habitat and barplot of the result with data Allen
}
