% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{make_permanova_barplot_allen_coral_mpa_with_all_species}
\alias{make_permanova_barplot_allen_coral_mpa_with_all_species}
\title{Barplot with all species : Permanova habitat + mpa}
\usage{
make_permanova_barplot_allen_coral_mpa_with_all_species(
  df,
  res1,
  res2,
  res3,
  res4,
  res5
)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Barplot with all species : Permanova habitat + mpa
}
