% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{make_transect_lines}
\alias{make_transect_lines}
\title{Make transect lines from points per sector}
\usage{
make_transect_lines(pts, sector)
}
\arguments{
\item{sector}{}
}
\value{

}
\description{
Make transect lines from points per sector
}
