% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{make_twoway_test_barplot_allen_coral_mpa}
\alias{make_twoway_test_barplot_allen_coral_mpa}
\title{Make two-way non parametric test (scheirerRayHare test) for density per mpa status and Allen coral habitat and barplot of the result}
\usage{
make_twoway_test_barplot_allen_coral_mpa(df, species)
}
\arguments{
\item{coral_attrib}{}
}
\value{

}
\description{
Make two-way non parametric test (scheirerRayHare test) for density per mpa status and Allen coral habitat and barplot of the result
}
