% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{make_twoway_test_barplot_allen_coral_mpa_with_all_species_class2_mpa_without_zero}
\alias{make_twoway_test_barplot_allen_coral_mpa_with_all_species_class2_mpa_without_zero}
\title{Barplot with all species : Make two-way non parametric test (scheirerRayHare test) for density per mpa status and Allen coral habitat and barplot of the result}
\usage{
make_twoway_test_barplot_allen_coral_mpa_with_all_species_class2_mpa_without_zero(
  df,
  species1,
  species2,
  species3,
  species4,
  species5
)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Barplot with all species : Make two-way non parametric test (scheirerRayHare test) for density per mpa status and Allen coral habitat and barplot of the result
}
