% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{map_allen_coral_poly}
\alias{map_allen_coral_poly}
\title{Make map of Allen coral geomorphology polygons benthic}
\usage{
map_allen_coral_poly(
  maplatlon,
  cor,
  extent,
  lon1,
  lon2,
  lat1,
  lat2,
  dist,
  offset_lon,
  offset_lat
)
}
\arguments{
\item{offset_lat}{}
}
\value{

}
\description{
Make map of Allen coral geomorphology polygons benthic
}
