% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{map_allen_coral_poly_coastline_with_open_sea}
\alias{map_allen_coral_poly_coastline_with_open_sea}
\title{Make map of Allen coral geomorphology polygons with added open sea category}
\usage{
map_allen_coral_poly_coastline_with_open_sea(
  maplatlon,
  pa,
  cor,
  extent,
  lon1,
  lon2,
  lat1,
  lat2
)
}
\arguments{
\item{pa}{}
}
\value{

}
\description{
Make map of Allen coral geomorphology polygons with added open sea category
}
