% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{map_allen_coral_poly_satellite}
\alias{map_allen_coral_poly_satellite}
\title{Make map of Allen coral geomorphology polygons with satellite view}
\usage{
map_allen_coral_poly_satellite(
  maplatlon,
  pa,
  cor,
  extent,
  lon1,
  lon2,
  lat1,
  lat2,
  dist,
  offset_lon,
  offset_lat
)
}
\arguments{
\item{dist}{}
}
\value{

}
\description{
Make map of Allen coral geomorphology polygons with satellite view
}
