% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{map_coral_cover}
\alias{map_coral_cover}
\title{Make map of coral cover raster}
\usage{
map_coral_cover(maplatlon, cor, extent)
}
\arguments{
\item{extent}{}
}
\value{

}
\description{
Make map of coral cover raster
}
