% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{map_coral_poly}
\alias{map_coral_poly}
\title{Make map of coral geomorphology polygons}
\usage{
map_coral_poly(maplatlon, cor, attrib, extent)
}
\arguments{
\item{extent}{}
}
\value{

}
\description{
Make map of coral geomorphology polygons
}
