% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_maps.R
\name{map_dens_per_grid_per_date_species_poe}
\alias{map_dens_per_grid_per_date_species_poe}
\title{Map density per grid cell for given species per date for POe (nb of observations / (length of tracks in meters * footprint_width in meters))}
\usage{
map_dens_per_grid_per_date_species_poe(
  maplatlonproj,
  polyobs,
  polytracks,
  footprintwidth,
  species
)
}
\arguments{
\item{maplatlonproj}{}
}
\value{

}
\description{
Map density per grid cell for given species per date for POe (nb of observations / (length of tracks in meters * footprint_width in meters))
}
