% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_maps.R
\name{map_dens_per_grid_species_poe_with_megafauna_image}
\alias{map_dens_per_grid_species_poe_with_megafauna_image}
\title{Map density per grid cell for given species for POe (nb of observations / (length of tracks in meters * footprint_width in meters))
with mpa overlaid and with megafauna image}
\usage{
map_dens_per_grid_species_poe_with_megafauna_image(
  maplatlonproj,
  polyobs,
  polytracks,
  footprintwidth,
  species,
  pa,
  img
)
}
\arguments{
\item{img}{}
}
\value{

}
\description{
Map density per grid cell for given species for POe (nb of observations / (length of tracks in meters * footprint_width in meters))
with mpa overlaid and with megafauna image
}
