% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{map_indiv_species_telemetry_separate_poe_on}
\alias{map_indiv_species_telemetry_separate_poe_on}
\title{Make map of telemetry with individual species observations with separate map per species for Poe on effort}
\usage{
map_indiv_species_telemetry_separate_poe_on(maplatlon, telem, telem_obs)
}
\arguments{
\item{just_poe}{}
}
\value{

}
\description{
Make map of telemetry with individual species observations with separate map per species for Poe on effort
}
