% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_maps.R
\name{map_nbr_survey_with_sightings_per_grid_image}
\alias{map_nbr_survey_with_sightings_per_grid_image}
\title{Map nbr survey with sightings per grid with image}
\usage{
map_nbr_survey_with_sightings_per_grid_image(
  maplatlonproj,
  polytracks,
  polygon,
  species,
  pa,
  img
)
}
\arguments{
\item{img}{}
}
\value{

}
\description{
Map nbr survey with sightings per grid with image
}
