% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_maps.R
\name{map_nbr_survey_with_sightings_per_grid_image_coastline}
\alias{map_nbr_survey_with_sightings_per_grid_image_coastline}
\title{Map nbr survey with sightings per grid with image with coastline for better view of image}
\usage{
map_nbr_survey_with_sightings_per_grid_image_coastline(
  polytracks,
  polygon,
  species,
  pa,
  img
)
}
\arguments{
\item{polytracks}{}
}
\value{

}
\description{
Map nbr survey with sightings per grid with image with coastline for better view of image
}
