% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{map_nc}
\alias{map_nc}
\title{Make open street map for nc area + australia}
\usage{
map_nc(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lon2}{}
}
\value{

}
\description{
Make open street map for nc area + australia
}
