% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{map_new_caledonia}
\alias{map_new_caledonia}
\title{Make map australia + nc : Figure 1}
\usage{
map_new_caledonia()
}
\arguments{
\item{lat2}{}
}
\value{



}
\description{
Make map australia + nc : Figure 1
}
\examples{
Make New caledonia map with Poé location

}
