% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_maps.R
\name{map_obs_per_grid_species_poe}
\alias{map_obs_per_grid_species_poe}
\title{Map number of observations per grid cell for given species for POe}
\usage{
map_obs_per_grid_species_poe(maplatlonproj, polygon, species)
}
\arguments{
\item{polygon}{}
}
\value{

}
\description{
Map number of observations per grid cell for given species for POe
}
