% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{map_telemetry_date_separate_on_off}
\alias{map_telemetry_date_separate_on_off}
\title{Make separate map of telemetry per date for all telemetry and telemetry on effort}
\usage{
map_telemetry_date_separate_on_off(
  maplatlon,
  telem,
  telem_on,
  extent,
  just_poe = FALSE
)
}
\arguments{
\item{extent}{}
}
\value{

}
\description{
Make separate map of telemetry per date for all telemetry and telemetry on effort
}
