% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{map_telemetry_date_separate_poe_on}
\alias{map_telemetry_date_separate_poe_on}
\title{Make separate map of telemetry per date for Poe on effort}
\usage{
map_telemetry_date_separate_poe_on(maplatlon, telem)
}
\arguments{
\item{extent}{}
}
\value{

}
\description{
Make separate map of telemetry per date for Poe on effort
}
