% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{map_telemetry_video_id_obs_separate_poe_on}
\alias{map_telemetry_video_id_obs_separate_poe_on}
\title{Make map of telemetry with one color per video_id with species observation for Poe on effort}
\usage{
map_telemetry_video_id_obs_separate_poe_on(
  maplatlon,
  telem,
  telem_obs,
  species
)
}
\arguments{
\item{telem}{}
}
\value{

}
\description{
Make map of telemetry with one color per video_id with species observation for Poe on effort
}
