% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_maps.R
\name{map_tracklen_per_grid_poe}
\alias{map_tracklen_per_grid_poe}
\title{Map track length per grid cell for Poe}
\usage{
map_tracklen_per_grid_poe(maplatlonproj, polytracks)
}
\arguments{
\item{polytracks}{}
}
\value{

}
\description{
Map track length per grid cell for Poe
}
