% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{map_transects}
\alias{map_transects}
\title{Make map of transect lines}
\usage{
map_transects(maplatlon, transects, extent, just_poe = FALSE)
}
\arguments{
\item{extent}{}
}
\value{

}
\description{
Make map of transect lines
}
