% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_maps.R
\name{map_transects_scalebar_mpa}
\alias{map_transects_scalebar_mpa}
\title{Make map of transect lines with scale bar and Poe mpa}
\usage{
map_transects_scalebar_mpa(
  maplatlon,
  transects,
  pa,
  extent,
  lat1,
  lon1,
  lat2,
  lon2,
  dist,
  offset_lon,
  offset_lat,
  just_poe = FALSE
)
}
\arguments{
\item{pa}{}
}
\value{

}
\description{
Make map of transect lines with scale bar and Poe mpa
}
