% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{permanova_barplot_allen_coral_mpa_with_all_species_habitat_mpa_without_zero}
\alias{permanova_barplot_allen_coral_mpa_with_all_species_habitat_mpa_without_zero}
\title{Barplot with all species : Permanova _habitat_mpa_without_zero}
\usage{
permanova_barplot_allen_coral_mpa_with_all_species_habitat_mpa_without_zero(
  df,
  species1,
  species2,
  species3,
  species4,
  species5
)
}
\arguments{
\item{species5}{}
}
\value{

}
\description{
Barplot with all species : Permanova _habitat_mpa_without_zero
}
