% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_densities.R
\name{permanova_barplot_allen_coral_mpa_with_all_species_mpa_habitat}
\alias{permanova_barplot_allen_coral_mpa_with_all_species_mpa_habitat}
\title{Barplot with all species : Permanova _mpa_habitat}
\usage{
permanova_barplot_allen_coral_mpa_with_all_species_mpa_habitat(
  df,
  species1,
  species2,
  species3,
  species4,
  species5
)
}
\arguments{
\item{species5}{}
}
\value{

}
\description{
Barplot with all species : Permanova _mpa_habitat
}
