% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{read_crop_and_convert_allen_coralnc_geomorphic}
\alias{read_crop_and_convert_allen_coralnc_geomorphic}
\title{Read and convert New Caledonia Allen coral shapefile geomorphic}
\usage{
read_crop_and_convert_allen_coralnc_geomorphic(lon1, lon2, lat2, lat1)
}
\arguments{
\item{lat1}{}
}
\value{

}
\description{
Read and convert New Caledonia Allen coral shapefile geomorphic
}
