% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{read_megafauna_image}
\alias{read_megafauna_image}
\title{Read megafauna image in png format (original black silhouette)}
\usage{
read_megafauna_image(species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Read megafauna image in png format (original black silhouette)
}
