% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{read_megafauna_image_grey}
\alias{read_megafauna_image_grey}
\title{Read megafauna image in png format (converted grey silhouette)}
\usage{
read_megafauna_image_grey(species)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Read megafauna image in png format (converted grey silhouette)
}
