% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_maps.R
\name{restrict_telem_dates}
\alias{restrict_telem_dates}
\title{restrict telemetry to dates}
\usage{
restrict_telem_dates(telem, dates)
}
\arguments{
\item{dates}{}
}
\value{

}
\description{
restrict telemetry to dates
}
