% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{select_obs_telemetry_poe}
\alias{select_obs_telemetry_poe}
\title{select observations for Poe}
\usage{
select_obs_telemetry_poe(telem_obs, ls_poe_vids)
}
\arguments{
\item{ls_poe_vids}{}
}
\value{

}
\description{
select observations for Poe
}
