% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{select_on_effort_poe}
\alias{select_on_effort_poe}
\title{Select on effort data (telemetry or observations) Poe}
\usage{
select_on_effort_poe(telem_poe, off_transit, off_loop)
}
\arguments{
\item{off}{}
}
\value{

}
\description{
Select on effort data (telemetry or observations) Poe
}
