% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_clean.R
\name{select_telemetry_in_buffer}
\alias{select_telemetry_in_buffer}
\title{select telemetry points in buffer}
\usage{
select_telemetry_in_buffer(telem_sp, buf)
}
\arguments{
\item{buf}{}
}
\value{

}
\description{
select telemetry points in buffer
}
